Fix  #2423: PingAsync() doesn't work if remote connection isn't authenticated
